newback/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==4.1.2
Django==4.2.16
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-timezone-field==7.0
django-unfold==0.38.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
isodate==0.6.1
jwcrypto==1.5.6
kombu==5.4.1
lxml==5.3.0
MetaTrader5==5.0.4803
msgpack==1.1.0
numpy==1.24.4
openpyxl==3.1.5
packaging==24.1
persiantools==4.2.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
psycopg==3.2.1
psycopg-binary==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.6.0
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2024.2
redis==5.0.8
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wfastcgi==3.0.0
zeep==4.2.1
zope.interface==7.0.3
